Updating task list works
